Remove custom signal handling from the parent process
Deferred handling of quit signals caused a racey bug where if the X
server or session was killed first and the parent immediately after, the
session would be restarted and the quit signal ignored.
Since only quit signals were handled, the default exit() should suffice
and the kernel should take care the cleanup.